refactor collate functions and trainer to return lists and splat them seperately for the 2d and the 3d model. Also finish implementing EGNNTorch

+++ trainer/metrics.py
@@ -111,7 +111,7 @@ class OGBEvaluator(nn.Module):
     def forward(self, preds, targets):
         if preds.shape[1] != self.evaluator.num_tasks:
             return torch.tensor(float('NaN'))
-        input_dict = {"y_true": targets, "y_pred": preds}
+        input_dict = {"y_true": targets.long(), "y_pred": preds}
         return torch.tensor(self.evaluator.eval(input_dict)[self.metric])
